2019-08-31 all commit

+++ ask-service/src/main/java/com/yiyn/ask/xcx/account/po/AccountWithDrawPo.java
@@ -10,9 +10,9 @@ public class AccountWithDrawPo extends BasePo {
 	 */
 	private static final long serialVersionUID = 1L;
 	private String account_id;
-	private double withdraw; //提现金额
-	private double withdraw_act; // 提现金额*30%
-	private double service_charge;//手续费
+	private BigDecimal withdraw; //提现金额
+	private BigDecimal withdraw_act; // 提现金额*30%
+	private BigDecimal service_charge;//手续费
 	private String withdraw_type; // 提现状态  1：微信  2：线下
 	private String status; // 提现状态
 
@@ -32,27 +32,27 @@ public class AccountWithDrawPo extends BasePo {
 		this.account_id = account_id;
 	}
 
-	public double getWithdraw() {
+	public BigDecimal getWithdraw() {
 		return withdraw;
 	}
 
-	public void setWithdraw(double withdraw) {
+	public void setWithdraw(BigDecimal withdraw) {
 		this.withdraw = withdraw;
 	}
 
-	public double getWithdraw_act() {
+	public BigDecimal getWithdraw_act() {
 		return withdraw_act;
 	}
 
-	public void setWithdraw_act(double withdraw_act) {
+	public void setWithdraw_act(BigDecimal withdraw_act) {
 		this.withdraw_act = withdraw_act;
 	}
 
-	public double getService_charge() {
+	public BigDecimal getService_charge() {
 		return service_charge;
 	}
 
-	public void setService_charge(double service_charge) {
+	public void setService_charge(BigDecimal service_charge) {
 		this.service_charge = service_charge;
 	}
